+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Benchmarks
 - HPOBench: Add rl, nasbench_101, nasbench_201, nasbench1shot1, and nas_hpo (#155, fix for benchmarks #158)
 
+## Improvements
+- Update/clean optimizer configs, update HEBO+Skopt a bit (#161)
+
 # v0.1.1
 
 ## Documentation